Cambio a bigInteger los campos de DPI

+++ database/migrations/2018_12_15_171421_create_table__p_e_r_s_o_n_a.php
@@ -17,7 +17,7 @@ class CreateTablePERSONA extends Migration
         Schema::create('PERSONA', function (Blueprint $table) {
             
             $table->engine = 'InnoDB';
-            $table->integer('cui')->unsigned();
+            $table->bigInteger('cui')->unsigned();
             $table->string('nombres',200);
             $table->string('apellidos',200);
             $table->integer('genero')->unsigned();
