+++ database/migrations/2018_12_15_173323_create_table__n_a_c_i_m_i_e_n_t_o.php
@@ -17,7 +17,7 @@ class CreateTableNACIMIENTO extends Migration
             
             $table->engine = 'InnoDB';
             $table->increments('acta_nac');
-            $table->integer('cui')->unsigned();
+            $table->bigInteger('cui')->unsigned();
             $table->integer('cui_padre')->unsigned();
             $table->integer('cui_madre')->unsigned();
             $table->integer('id_muni')->unsigned();
